Updated GridManager functionality and project configurations
Updated `GridManager.cs` to introduce randomness in the cell update process of the grid, generating more dynamic visual patterns in the Game of Life implementation. Made `gridWidth` and `gridHeight` properties to be set dynamically based on the viewport size. Introduced a `ToggleCell` function to capsulate the logic of cell state update. Added random color generating for each cell.

`.gitignore` file is updated to ignore `.fake` files.

`project.godot` is updated to set `node_2d.tscn` as the main scene to run.