Add option to draw dead cells and enhance neighbor counting
Added a new public boolean variable "drawDeadCell" which when enabled, dead cells in the grid are drawn. This improvement provides visual cues aiding in debugging the grid functions.

The code for counting alive neighbors is updated to be more efficient and readable. The old method of using nested loops is replaced by using two fixed arrays for the X and Y offsets of the 8 neighboring cells. This method defines explicitly the neighbours and makes code more understandable.

The neighbour counting now supports edges wrap around to create a toroidal array effect. This change can affect the dynamics of the simulation, especially for spaceships that can now cross the grid boundaries.

Formatting changes also made for improved readability.