Add functionality to draw specific patterns on grid
This commit adds the ability to draw two specific patterns, the "Glider" and "Gosper's Glider Gun", on the grid by pressing Key1 and Key2 respectively. It also introduces the logic to draw these patterns at the mouse position on the grid. A new Pattern class was created along with a DrawPattern method to support this. The goal is to enrich user interaction and showcase the potential of Conway's Game of Life.