"Add matrix manipulation class for cleaner code and new features"
Extracted matrix-related functions into a separate MatrixManipulation class to clean up GridManager and increase modularity. Incorporated rewind and step forward functions in order to enhance game control. A SaveState method was also added to allow for the preservation of game states, which makes the rewind and step forward features possible.