Relocate QueueRedraw() call to optimize rendering
QueueRedraw() method call was moved from the _Process() method in GridManager.cs, the SaveState() method in GridController.cs, and the ProcessTurn() and Rewind() methods in Grid.cs to the _Draw() method in GridManager.cs. This serves to optimize our rendering process by preventing unnecessary redraw calls that could be made when the grid state isn't changing, thus enhancing the performance of our application.