Refactored GridManager to handle grid rendering
With this commit, GridRenderer.cs was deleted and its grid drawing logic was moved into the new class GameGridHandler.cs. The refactoring was done to comply with Single Responsibility Principle. The grid management and grid drawing functions are more integral to each other and placing them in the same class thus makes the code easier to understand and maintain.