Refactor game grid handler and integrate pause feature
The matrix manipulation instance initialization was moved to the top for better visibility. Now, the process of the controller only continues, if the game is not paused. This prevent the controller from redrawing while the game is in pause state. Also, the input for the system is not directly managed by the GameGridHandler anymore, which further decouples components. The Implementation for '_grid.StepForward()' in 'InputManager.cs' has been added. Simultaneously 'GridManager.cs' is replaced with 'GameGridHandler.cs'. This makes the architecture clearer and independent.