Refactor GameGridHandler for readability and organization
In GameGridHandler.cs, the initializations of the various components were moved into a separate InitComponents function to isolate their initializations and make them more modular. This improves readability and reduces clutter in the _Ready class. Similarly, the addition of these components to the scene was moved into a separate AddComponentsToScene function. DrawCells() method is refactored to avoid redundancy, while DrawDebugLinesIfEnabled() to improve readability.

We also added a guard clause in the _Draw function to prevent errors when accessing states outside of the acceptable range, making the code more resilient.