Refactor code and enhance game interface display
This commit includes multiple changes that seek to enhance readability and improve the player's interaction with the game. On code level, unnecessary usings were removed from the GameGridHandler.cs script and condition checks in Grid.cs have been compacted for simplicity.

The most notable change includes updates made to node_2d.tscn scene file to redesign game's interface to be more informative by adding rules, controls, and mouse interactions to the game display text. The objective is to provide clear instructions on how players can interact with the Conway's Game of Life game scene. This is an important modification to boost the UX and game accessibility.