Refactored code to comply with C# 10 syntax
In this commit, we have updated the syntax of two files, 'Cell.cs' and 'InputManager.cs', to align with the changes in C# 10. The changes are mainly about the new syntax for namespaces that allows for a more compact hierarchy and the removal of unnecessary using statements. Specifically, we've brought in the semicolon-separated, multiline namespaces syntax into the codebase to enhance readability and ease of structuring. We've also removed a redundant `using System.Diagnostics.CodeAnalysis;` line since it was unused. This refactoring is aimed to make our code more maintainable and up-to-date with modern C# practices.