Add copy pattern functionality and update debug outputs
Added a new `CopyPattern` function to the `PatternCreator` class for feature updates requiring pattern duplication. Renamed `PrintAllNodes` method in 'GameGridHandler' to `DebugPrintAllNodes` to clearly signify that it is a debugging tool. Changed the method `DrawBox's` modifier from public to private as it should not be exposed for external use. Furthermore, `CopyPattern` is now called during `MirrorAndShift` operation in `Grid.cs` to enhance grid manipulation capabilities.