"Added 'State' property to Cell and implemented Brian's Brain cellular automaton rules
In addition to Conway's Game of Life rules, Brian's Brain rules were introduced. Therefore, a 'State' property was added to the Cell class to handle different cell states ('Alive', 'Dying', 'Dead'). The grid initialization was also updated to set the initial state of all cells to 'Alive'. The 'GetNewCell' function was renamed to 'GetNewCellByConwaysRules' for clarity, and a new function 'GetNewCellByBriansBrainRules' was created to generate cells by the new rules."