Add 'Rabit01' pattern to PatternCreator and InputManager
A new pattern called 'Rabit01' has been added to PatternCreator class. Accordingly,
an input key binding for 'Rabit01' pattern has been added to the InputManager class which allows you to draw the 'Rabit01' pattern on the grid. This addition extends the functionality of the game by introducing more variation in patterns.