Refactor code to improve readability and maintainability
In this commit, a substantial chunk of code for creating an initial grid and patterns in the Grid.cs file has been moved into a new method GetInitialGrid() in PatternCreator.cs. This contributes to making the code easier to read and maintain and adheres to the DRY principle of software development.

Additionally, a typo in InputManager.cs has been fixed (Rabit01 to Rabbit01) and some unnecessary white spaces in Grid.cs have been removed. The order of variable declarations in Grid.cs has also been re-formatted. This was done to enhance readability and adhere to standard naming conventions.