Move CopyPattern method from PatternCreator to InputManager"
In this commit, the CopyPattern() method is moved from the PatternCreator.cs file to the InputManager.cs file. This is done to consolidate all input-related methods within the same class, improving code organisation and readability. Additionally, an unused call to the CopyPattern() method in Grid.cs is removed. A new method for mirroring the matrix is also added in MatrixManipulation.cs.