Add bug fix comment and refactor ClearGrid method in MatrixManipulation
In InputManager.cs, a comment was added to address an existing bug: Mirrored cells not being drawn correctly and exiting cells being erroneously removed.

In node_2d.tscn, added unique property to GenCounter's name to avoid any future conflict.

In MatrixManipulation.cs, the method ClearGrid was refactored to prevent potential NullReferenceException, by checking for null before accessing the Grid object. This will improve code stability.
We have also shifted from clearing all cells of the grid one-by-one to creating an empty grid in one go. This change was made to improve the efficiency of the method and reduce complexity.