Refactored grid update logic and added color randomization
A considerable change was made to GridController.cs as the general flow of updating the grid was revised to simplify and improve accuracy. The GetNewCellByConwaysRules method is now used that accounts for cases of live or dead cells and adjusts their state for the next generation based on Conway's rules of Game of Life. The new process improved managing grid updates concerning the rules of neighbours and edge cases.

Additionally, the ability to randomize the colors of the cells was added to enhance visualization. This feature can be toggled on and off and is implemented in Grid.cs and CellPool.cs. Grid now keeps track of whether random colors should be used and CellPool sets a random color for each cell when randomization is on.

Lastly, minor improvements were made across multiple files for better code consistency and clarity. In PatternCreator, the variables were renamed to use camelCase and in GameGridHandler, the initialPoolSize was increased to manage the larger grid more efficiently. A new input for toggling color randomization was added to InputManager. Furthermore, cells are now correctly released back to the pool in Grid when they die.

These changes result in richer visualization options for the user and more efficient updates, improving the overall performance and user experience.