Refactor code and preserve color of living cells
Changed the data type of 'stateOfCell' variable from 'bool' to 'var' to allow greater flexibility for future changes. Also, added a condition to preserve the color of current living cells in the new generation, enhancing the visual continuity and user's experience. The color only changes when a new cell is born or revived.